addition in hl teaser view and issue solving

+++ service-loans/src/main/java/com/capitaworld/service/loans/service/teaser/primaryview/impl/PlTeaserViewServiceImpl.java
@@ -298,7 +298,7 @@ public class PlTeaserViewServiceImpl implements PlTeaserViewService {
 				
 				plTeaserViewResponse.setCity(CommonDocumentUtils.getCity(plRetailApplicantRequest.getAddressCity(), oneFormClient));
 				plTeaserViewResponse.setState(CommonDocumentUtils.getState(plRetailApplicantRequest.getAddressState(), oneFormClient));
-				plTeaserViewResponse.setCountry(CommonDocumentUtils.getState(plRetailApplicantRequest.getAddressCountry(), oneFormClient));
+				plTeaserViewResponse.setCountry(CommonDocumentUtils.getState(plRetailApplicantRequest.getAddressCountry().longValue(), oneFormClient));
 				
 				// address
 				
@@ -856,7 +856,7 @@ public class PlTeaserViewServiceImpl implements PlTeaserViewService {
 				//citetailApplicantResponse.setry,State,country
 				plTeaserViewResponse.setCity(CommonDocumentUtils.getCity(plRetailApplicantRequest.getAddressCity(), oneFormClient));
 				plTeaserViewResponse.setState(CommonDocumentUtils.getState(plRetailApplicantRequest.getAddressState(), oneFormClient));
-				plTeaserViewResponse.setCountry(CommonDocumentUtils.getState(plRetailApplicantRequest.getAddressCountry(), oneFormClient));
+				plTeaserViewResponse.setCountry(CommonDocumentUtils.getState(plRetailApplicantRequest.getAddressCountry().longValue(), oneFormClient));
 				
 				// address
 				try {
